firewall network config-ed. Working on ubuntu network config script

+++ lxc-uninstall.sh
@@ -1,9 +1,15 @@
+user=$(whoami)
+if [ $user != 'root' ]; then
+  echo "El programa ha de ejecutarse como root"
+  exit
+fi
+
 echo "Borrando contenedores..."
 lxc delete firewall --force
 lxc delete loadbalancer --force
-lxc delete webserver0 --force
 lxc delete webserver1 --force
 lxc delete webserver2 --force
+lxc delete webserver3 --force
 lxc delete storage0 --force
 lxc delete storage1 --force
 lxc delete storage2 --force
